Exception catch, fixed hardcoded user
I was running into some exceptions with foreign keys (this ended up
being because I still had feeds updating), so I added a try/catch to try
and see what was going on.  I also changed the hardcoded user in
UpdateSeries() to use the psql user that was passed in.